refactor: service interface and imple
